[Parser] fix Onnx dtype issuse (CB-6835)

+++ AIPUBuilder/Parser/ops/onnx_ops/math_ops.py
@@ -696,7 +696,7 @@ class HardSigmoidOp(LayoutUnawareOp, BaseActivationOp, OnnxOp):
         inputs = self.get_input_tensors()
         out_tensor = np.maximum(self.clip_min,
                                 np.minimum(self.clip_max,
-                                           self.alpha * inputs[0] + self.beta))
+                                           self.alpha * inputs[0] + self.beta)).astype(inputs[0].dtype)
         self.set_out_tensor(out_tensor)
